backend/requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.3.0
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe>=2.1.0
Werkzeug==3.1.3
gunicorn==22.0.0
eventlet==0.33.3
solana==0.29.1
solders==0.14.0
base58==2.1.1
httpx==0.27.0
spl-token==0.4.1
typing_extensions==4.14.0
